### 2023-07-31 #5712 @editorialbot
[PRE REVIEW]: `pgmuvi`: Quick and easy Gaussian Process Regression for multi-wavelength astronomical timeseries
**Submitting author:** @pscicluna (<a href="http://orcid.org/0000-0002-1161-3756">Peter Scicluna</a>)
**Repository:** https://github.com/ICSM/pgmuvi
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
