### 2023-08-04 9c561ea
Create CODE_OF_CONDUCT.md
