### 2023-08-04 8148331
Implement issue templates
Trying out bug report and feature request templates for the community to work with.
