+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,23 @@
+---
+name: Feature request
+about: Suggest an idea for this project
+title: "[Feature Request]"
+labels: ''
+assignees: ''
+
+---
+
+**Is your feature request related to a problem? Please describe.**
+A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+
+**Describe the solution you'd like**
+A clear and concise description of what you want to happen. Ideally, include an example, or refer to another package, gist, or paper.
+
+**Describe alternatives you've considered**
+A clear and concise description of any alternative solutions or features you've considered.
+
+**Optional: propose implementation**
+If you've already considered if, please propose a possible approach to implementing the feature. If not, that can be discussed later.
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
