+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,35 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: "[BUG]"
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behaviour:
+
+
+**Expected behaviour**
+A clear and concise description of what you expected to happen.
+
+**Environment details**
+Please provide the operating system name and version. 
+Please either attach the output of `pip freeze` or provide the versions of the following packages:
+python = 
+numpy = 
+scipy = 
+torch = 
+xarray = 
+pyro = 
+gpytorch = 
+matplotlib = 
+seaborn = 
+arviz = 
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
