### 2023-08-18 b240c4e
Try casting to array for Table creation
