### 2023-08-19 f4c5a89
More robust conversion to array for output.
