### 2023-08-20 bdf3571
Force casting list(*tensor) to tensor before casting to ndarray
