### 2023-08-20 f55b5e8
Make sure that each column only has one entry ...
even if it is sometimes a long array
