+++ pgmuvi/lightcurve.py
@@ -2004,37 +2004,37 @@ class Lightcurve(object):
         """
         from astropy.table import Table
         t = Table()
-        t['x'] = np.asarray(self.xdata)
-        t['y'] = np.asarray(self.ydata)
+        t['x'] = [np.asarray(self.xdata)]
+        t['y'] = [np.asarray(self.ydata)]
         if hasattr(self, 'yerr'):
-            t['yerr'] = np.asarray(self.yerr)
+            t['yerr'] = [np.asarray(self.yerr)]
         if self.__FITTED_MCMC or self.__FITTED_MAP:
             # These outputs can only be produced if a fit has been run.
             periods, weights, scales = self.get_periods()
-            t['period'] = np.asarray(periods)
+            t['period'] = [np.asarray(periods)]
             try:
-                t['weights'] = np.asarray(weights)
+                t['weights'] = [np.asarray(weights)]
             except RuntimeError:
-                t['weights'] = torch.as_tensor(weights).detach().numpy()
+                t['weights'] = [torch.as_tensor(weights).detach().numpy()]
             try:
-                t['scales'] = np.asarray(scales)
+                t['scales'] = [np.asarray(scales)]
             except RuntimeError:
-                t['scales'] = torch.as_tensor(scales).detach().numpy()
+                t['scales'] = [torch.as_tensor(scales).detach().numpy()]
             for key, value in self.results.items():
                 try:
-                    t[key] = np.asarray(value)
+                    t[key] = [np.asarray(value)]
                 except RuntimeError:
-                    t[key] = torch.as_tensor(value).detach().numpy()
+                    t[key] = [torch.as_tensor(value).detach().numpy()]
             if self.__FITTED_MAP:
                 # Loss isn't relevant for MCMC, I think
-                t['loss'] = np.asarray(self.results['loss'])
+                t['loss'] = [np.asarray(self.results['loss'])]
             # Now we want the model predictions for the input times:
             if self.__FITTED_MAP:
                 with torch.no_grad():
                     observed_pred = self.likelihood(self.model(self._xdata_raw))
-                    t['y_pred_mean'] = np.asarray(observed_pred.mean)
-                    t['y_pred_lower'] = np.asarray(observed_pred.confidence_region()[0])
-                    t['y_pred_upper'] = np.asarray(observed_pred.confidence_region()[1])
+                    t['y_pred_mean'] = [np.asarray(observed_pred.mean)]
+                    t['y_pred_lower'] = [np.asarray(observed_pred.confidence_region()[0])]
+                    t['y_pred_upper'] = [np.asarray(observed_pred.confidence_region()[1])]
             elif self.__FITTED_MCMC:
                 raise NotImplementedError("MCMC predictions not yet implemented")
                 # with torch.no_grad():
