### 2023-08-20 15cdb7c
Excluding some lines so Ruff doesn't complain
