### 2023-08-21 c53e55a
Ensure plot() doesn't need to be called first ...
... and that the model is predicted both on the observed points and on a regular grid.
