### 2023-08-22 eae210c
Add test in test.py
