### 2023-08-22 2c82778
Add ndim() test in lighcurve class
