### 2023-08-24 56f19f8
Add test for xdata method
