### 2023-08-24 8fdda6c
Add test for ydata method
