### 2023-08-24 a5a4983
Add test for yerr
