### 2023-08-24 fbba3e3
Add test for xdata, ydata, yerr
