### 2023-08-24 b0d52f7
Updating when actions are run
We want tests and linting to run on every push, but the paper builds only need to run on the main branch.
