+++ .github/workflows/ruff.yml
@@ -1,10 +1,6 @@
 name: Ruff
 
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
+on: [push]
 
 jobs:
   ruff:
