+++ .github/workflows/paper-draft.yml
@@ -1,4 +1,9 @@
-on: [push]
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
 
 jobs:
   paper:
