+++ .github/workflows/paper-draft-tex-and-pdf.yml
@@ -10,9 +10,9 @@ name: Build JOSS paper and tex
 
 on:
   push:
-    branches:
-    - main
+    branches: [ main ]
   pull_request:
+    branches: [ main ]
 
 jobs:
   paper:
