+++ .github/workflows/test.yml
@@ -1,11 +1,7 @@
 
 name: Test
 
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
+on: [push]
 
 jobs:
   test:
