### 2023-08-24 e3181cb
Start by inhereting from `Module`
