### 2023-08-24 d83ac40
Registering buffers in Transformers
Now they should be good to go for moving to devices.
