### 2023-08-24 f0484f7
registering buffers in Lightcurve
I think I got everything?
