### 2023-08-24 5f845a5
Make sure `super().__init__() is called
`nn.Module.__init__()` does some important setup steps to make everthing work as intended - now that our objects are subclasses of that, we need to make sure we use `super()` to init the instances correctly;
