### 2023-08-24 aaa9679
Testing inline buffer registration
