### 2023-08-24 30e04bf
Make sure Tensors are on CPU before casting to numpy
