### 2023-08-24 dc121be
Making sure things are on CPU before numpy()
