### 2023-08-24 c433505
Moving more tensors to CPU for numpy
