### 2023-08-24 18d2973
Move explicit device tracking
