### 2023-08-24 037dc01
Trying to debug MCMC CUDA runs
