### 2023-08-24 bc0ba65
Does moving to GPU before setting priors make a difference?
