### 2023-08-24 925b7de
Trying solution from https://github.com/cornellius-gp/gpytorch/issues/1142
