### 2023-08-24 73d7b0f
Testing a different `pyro_model`
