### 2023-08-24 bb0679b
Switching back to the old pyro_model, but no dice
