### 2023-08-25 ecdf0d4
Trying to find non-cuda tensors
