### 2023-08-25 9fdc2b6
checking after priors are set
