+++ pgmuvi/lightcurve.py
@@ -1153,15 +1153,6 @@ class Lightcurve(torch.nn.Module):
 
         if cuda:
             self.cuda()
-            for key in self.state_dict().keys():
-                print(key)
-                try:
-                    print(self.state_dict()[key].device)
-                except AttributeError:
-                    pass
-                #self.state_dict()[key] = self.state_dict()[key].cuda()
-            for param_name, param in self.model.named_parameters():
-                print(f'Parameter name: {param_name:42} value = {param.data}, device = {param.data.device}')  # noqa: E501
 
         if not self.__PRIORS_SET:
             self.set_default_priors()
@@ -1198,6 +1189,17 @@ class Lightcurve(torch.nn.Module):
                              num_chains=num_chains,
                              disable_progbar=disable_progbar)
         import linear_operator.utils.errors as linear
+        for key in self.state_dict().keys():
+            print(key)
+            try:
+                print(self.state_dict()[key].device)
+            except AttributeError:
+                pass
+            #self.state_dict()[key] = self.state_dict()[key].cuda()
+        for param_name, param in self.model.named_parameters():
+            print(f'Parameter name: {param_name:42} value = {param.data}, device = {param.data.device}')  # noqa: E501
+
+
         try:
             if cuda:
                 self.mcmc_run.run(self._xdata_transformed.cuda(),
