### 2023-08-25 678b5de
Also test right before where the error occurs
