### 2023-08-25 ef6c1a8
Simplifying printout of lookup dict
