### 2023-08-25 a00cb1c
Trying to implement more agressive cuda and cpu fucntions...
Who knows if this will work.
