### 2023-09-01 876ece4
created lightcurve.to_table method
Code from the write_votable method was moved into the to_table method. The write_votable method now calls to_table, and only writes the result to file.
