### 2023-09-01 37deb58
Update lightcurve.py
