### 2023-09-01 47bc847
squeeze() all the tensors in to_table
The to_table method now squeezes all the tensors before appending them to the astropy table
