### 2023-09-01 8d39891
squeeze removed from to_table
The to_table method is now only a simple copy of code from write_votable
