### 2023-09-05 d2e2666
Switching torch.nn.Module for gpytorch.Module
since that provides mode gpytorch-specific functionality which we will otherwise be missing.
