### 2023-09-11 f1dea85
Create _ensure_tensor() method and apply to xdata and ydata setter
