### 2023-09-12 b9e7185
Add _ensure_dim() method
