### 2023-11-05 cbd2345
Add installation instructions to README.md
