### 2024-02-23 #5876 @adonath
Just to confirm: the issues related to JOSS review are still open: https://github.com/ICSM/pgmuvi/issues?q=is%3Aopen+is%3Aissue+label%3A%22JOSS+review%22

@pscicluna Please address the review comments in a timely manner. Thanks!
