### 2024-09-09 1eeacdb
Update paper.md
whic --> which
