### 2024-09-09 64f65dc
Update paper.md
Rephrased first sentence
