### 2024-09-12 caada66
Ignoring some long lines in Ruff
