+++ pgmuvi/lightcurve.py
@@ -452,7 +452,7 @@ class Lightcurve(gpytorch.Module):
             self.likelihood = gpytorch.likelihoods.FixedNoiseGaussianLikelihood(self._yerr_transformed)  # noqa: E501
         elif hasattr(self, '_yerr_transformed') and likelihood == "learn":
             self.likelihood = gpytorch.likelihoods.FixedNoiseGaussianLikelihood(self._yerr_transformed,  # noqa: E501
-                                                                                learn_additional_noise=True)  # noqa: E501
+                                                                                learn_additional_noise=True)
         elif likelihood == "learn":
             self.likelihood = gpytorch.likelihoods.GaussianLikelihood(learn_additional_noise=True)  # noqa: E501
         elif "Constraint" in [t.__name__ for t in type(likelihood).__mro__]:
@@ -701,8 +701,8 @@ class Lightcurve(gpytorch.Module):
                                                                     apply_to=[0],
                                                                     shift=False))
                             constraint[key].lower_bound = torch.tensor(1./self.xtransform.transform(1./constraint[key].lower_bound,  # noqa: E501
-                                                                                                    apply_to=[0],  # noqa: E501
-                                                                                                    shift=False).item())  # noqa: E501
+                                                                                                    apply_to=[0],
+                                                                                                    shift=False).item())
                             if debug:
                                 print(constraint[key].lower_bound)
                                 print(constraint[key])
@@ -710,8 +710,8 @@ class Lightcurve(gpytorch.Module):
                             not in [torch.tensor(0), torch.tensor(torch.inf)]):
                             # we need to transform the upper bound
                             constraint[key].upper_bound = torch.tensor(1./self.xtransform.transform(1./constraint[key].upper_bound,  # noqa: E501
-                                                                                                    apply_to=[0],  # noqa: E501
-                                                                                                    shift=False).item())  # noqa: E501
+                                                                                                    apply_to=[0],
+                                                                                                    shift=False).item())
                             if debug:
                                 print(constraint[key].upper_bound)
                                 print(constraint[key])
@@ -978,7 +978,7 @@ class Lightcurve(gpytorch.Module):
 
         # for key in self._model_pars:
         #     with contextlib.suppress(AttributeError):
-        #         self._model_pars[key]['param'] = self._model_pars[key]['param'].cuda(device=device)
+        #         self._model_pars[key]['param'] = self._model_pars[key]['param'].cuda(device=device) # noqa: E501
         # try:
         #     self.model.cuda()
         #     self.likelihood.cuda()
@@ -1218,15 +1218,12 @@ class Lightcurve(gpytorch.Module):
         #     return y
 
         def pyro_model(x, y):
-            with (gpytorch.settings.fast_computations(False, False, False),
-                  gpytorch.settings.max_cg_iterations(max_cg_iterations)):
+            with (gpytorch.settings.fast_computations(False, False, False), gpytorch.settings.max_cg_iterations(max_cg_iterations)):  # noqa: E501
                 for key in self.state_dict().keys():
                     print(key)
-                    try:
+                    with contextlib.suppress(AttributeError):
                         print(self.state_dict()[key].device)
-                    except AttributeError:
-                        pass
-                    #self.state_dict()[key] = self.state_dict()[key].cuda()
+                                #self.state_dict()[key] = self.state_dict()[key].cuda()
[…]
