### 2024-09-12 3b4da84
Simplifying `dict_walk_generator`
It now uses unpacking, and `yield from` instead of loops
