+++ pgmuvi/lightcurve.py
@@ -52,16 +52,14 @@ def dict_walk_generator(indict, pre=None):
     if isinstance(indict, dict):
         for key, value in indict.items():
             if isinstance(value, dict):
-                for d in dict_walk_generator(value, pre + [key]):
-                    yield d
-            elif isinstance(value, list) or isinstance(value, tuple):
+                yield from dict_walk_generator(value, [*pre, key])
+            elif isinstance(value, (list, tuple)):
                 for v in value:
-                    for d in dict_walk_generator(v, pre + [key]):
-                        yield d
+                    yield from dict_walk_generator(v, [*pre, key])
             else:
-                yield pre + [key, value]
+                yield [*pre, key, value]
     else:
-        yield pre + [indict]
+        yield [*pre, indict]
 
 
 class Transformer(torch.nn.Module):
