### 2024-09-12 41f16b8
ensure that `psd` has shape consistent with `means` and `scales`
Applied `unsqueeze` if `psd` has fewer dimensions than `means`, so that the summation is performed along the correct axis.
