+++ pgmuvi/lightcurve.py
@@ -1892,6 +1892,8 @@ class Lightcurve(gpytorch.Module):
         psd1 = norm.log_prob(freq.unsqueeze(-1))
         psd2 = norm.log_prob(-freq.unsqueeze(-1))
         psd = torch.log(0.5 * (torch.exp(psd1) + torch.exp(psd2)))
+        if len(psd.shape) < len(means.shape):
+          psd = psd.unsqueeze(-1)
         if debug:
             print(psd.shape)
         try:
