### 2024-09-16 895e0a3
Numerically more stable PSD computation
The new version is numerically more stable in case psd2 is smaller than psd1.
