### 2024-09-16 ef97c6f
applying torch.log to Tensor([0.5])
