+++ pgmuvi/lightcurve.py
@@ -1891,7 +1891,7 @@ class Lightcurve(gpytorch.Module):
             print(norm)
         psd1 = norm.log_prob(freq.unsqueeze(-1))
         psd2 = norm.log_prob(-freq.unsqueeze(-1))
-        psd = torch.log(0.5) + psd1 + torch.log(1.0 + torch.exp(psd2 - psd1))
+        psd = torch.log(torch.Tensor([0.5])) + psd1 + torch.log(1.0 + torch.exp(psd2 - psd1))
         # psd = torch.log(0.5 * (torch.exp(psd1) + torch.exp(psd2)))
         if len(psd.shape) < len(means.shape):
           psd = psd.unsqueeze(-1)
