### 2024-09-16 fde3ae3
psd marginalisation fix
The PSD is correctly marginalised over the Fourier dual variable axis (-1, which has length 1 for a 1D light curve) instead of applying squeeze/unsqueeze.
This is in preparation for implementing 2D PSDs.
