### 2024-09-16 83ef566
Correctly broadcasting `weights` when calling `logsumexp`
