### 2024-09-16 77341ce
`except` code for `logsumexp` in `compute_psd` fixed
