### 2024-09-16 53f0993
Avoiding default constraints for 2D data
Since the assumption is that you cannot constrain a period which is longer than the time baseline of the data, but for other dimensions we may want to include long periods (e.g. something like an exponential decay in wavelength) quite often, we should not set the default constraints on the mixture means.

Also, trying to make this important information print in red.
