### 2024-09-16 45a0522
PSDs computed for lightcurve_2d objects
compute_psd upgraded for lightcurve_2d objects
