### 2024-09-16 10c0461
psd_compute compares freq to self.ndim
Checks whether the properties of freq are consistent with the dimensionality of the instantiated light curve object.
