### 2024-09-17 34f4b36
Check each input for NaNs
