### 2024-09-18 b348aed
Create `from_table` method
Read in (1d or 2d) light curve data from a VOTable. Information in table must be compatible with output from `to_table` method.
