+++ pgmuvi/lightcurve.py
@@ -328,6 +328,41 @@ class Lightcurve(gpytorch.Module):
         self.__FITTED_MAP = False
         self.__FITTED_MCMC = False
 
+    @classmethod
+    def from_table(cls, VOTable, **kwargs):
+        """Instantiate a Lightcurve object with
+        data read in from a VOTable.
+
+        Parameters
+        ----------
+        VOTable : str
+            Name (with extension) of the VOTable containing input data
+        kwargs:
+            Arguments to be passed to the Lightcurve call
+        
+        Returns
+        ----------
+        Lightcurve object
+        """
+        from astropy.table import Table
+        data = Table.read(VOTable, format='votable')
+        c = data.colnames
+        if ('x' not in c) or ('y' not in c):
+            raise ValueError('Table must have columns x, y, and yerr')
+        ndim = data['x'].squeeze().shape
+        x = torch.Tensor(data['x']).squeeze()
+        y = torch.Tensor(data['y']).squeeze()
+        if (len(ndim) == 1) or (len(ndim) == 2):
+            if 'yerr' not in c:
+                yerr = None
+            else:
+                yerr = torch.Tensor(data['yerr']).squeeze()
+        else:
+            mesg = "Column 'x' must have shape (1, nsamples) or (1, 2, nsamples)"
+            raise ValueError(mesg)
+    
+        return cls(x, y, yerr, **kwargs)
+    
     @property
     def ndim(self):
         return self.xdata.shape[-1] if self.xdata.dim() > 1 else 1
