### 2024-09-19 849969d
`yerr` mention removed from `from_table` error message
