### 2024-09-19 17da9de
Added keywords for column names, separate checks
`from_table` now accepts strings containing the names of the columns in the table containing the x, y, and (optionally) yerr columns. In addition, separate checks are now done for the existence of the x and y columns. Error message no longer mentions the optional yerr column.
