### 2024-09-19 47b5ea3
Correct definition of `ndim`
Number of dimensions is now defined as `len(...shape())` instead of `...shape()`.
