### 2024-09-19 6810c20
Added `table_name` and `format` arguments
`from_table` no longer requires VOTables; any supported format can be read in using `astropy.table.Table.read`.
