### 2024-09-19 cd17ba9
`from_table` accepts astropy.table.Table or str file name
