### 2024-09-19 9ec814e
`from_table` now accepts pathlib.Path instance
