### 2024-09-19 75a1e53
Combined `str` or `pathlib.Path` check for `tab`
