### 2024-09-19 68e0cc2
If `tab` is an `astropy.table.Table`, set `data=tab`
