### 2024-09-19 72e1ce8
Not casting `Path` object to `str`
