+++ pgmuvi/lightcurve.py
@@ -358,7 +358,7 @@ class Lightcurve(gpytorch.Module):
         from pathlib import Path
         from astropy.table import Table
         if isinstance(tab, str) or isinstance(tab, Path):
-            data = Table.read(str(tab), format=format)
+            data = Table.read(tab, format=format)
         elif isinstance(tab, astropy.table.Table):
             data = tab
         else:
