### 2024-09-21 2615f26
Starting to develop (hopefully) better models
