### 2024-09-23 94e184f
Added `fit_LS` method
First attempt at method to compute the Lomb-Scargle periodogram
