### 2024-09-23 72a6f2a
Updated docstring for `fit_LS`
