### 2024-09-23 aa5ef39
Return freq and PSD when `only_freq` set
