### 2024-09-23 2fc3a12
Update to `fit_LS`
reversed the sign in the following check to correctly identify true peaks:

if fap_max < single_threshold:
changed to
if fap_max > single_threshold:

On the subsequent line, removed hardcoding of the returned mask:

                return freq[peaks[:num_peaks]], np.array([True] * num_peaks)
