### 2024-09-23 6b9b516
Update `fit_LS`
Should return an array of `False` (and not `True`) when `fap_max > single_threshold`.
