+++ pgmuvi/lightcurve.py
@@ -1206,7 +1206,7 @@ are not yet implemented for 2D data
             # Calculate the false alarm probability for the highest peak
             fap_max = LS.false_alarm_probability(power.max())
             if fap_max > single_threshold:
-                return freq[peaks[:num_peaks]], np.array([True] * num_peaks)
+                return freq[peaks[:num_peaks]], np.array([False] * num_peaks)
             # Calculate the false alarm probability for each peak independently
             fap_single = LS.false_alarm_probability(power[peaks],
                                                     method='single')
