### 2024-09-23 a841547
`fit_LS` now returns `torch.Tensor` objects
