### 2024-09-23 f754843
`plot` method `ylim` set using data limits
`ylim = [-3, 3]` is very restrictive, setting the axis range with the data at least allows the user to quickly inspect the fit as soon as it is computed with minimal adjustment.
