### 2024-09-24 cf640f0
Corrected check for astropy.table.Table object (minor)
