### 2024-10-03 #5876 @warrickball
Hi @pscicluna, as we haven't heard from you despite multiple attempts on different channels, even though there's been activity in the repo (though none obviously towards this JOSS review), we'll close this issue and regard the submission as withdrawn in a week's time.
